@@ -1,5 +1,4 @@
-﻿@model ISM6225FinalProject.Models.Places
-<h1>About Team 5</h1>
+﻿<h1>About Team 5</h1>
 <hr />
 <p>
     <table>
