@@ -42,7 +42,7 @@
 
                 </div>
             </td>
-    <td width="75%" valign="top">
+    <td width="80%" valign="top">
         <div class="container" style="width:60%;">
             <h2>Some Images</h2>
             <div id="myCarousel" class="carousel slide" data-ride="carousel">
@@ -60,31 +60,51 @@
                 <!-- Wrapper for slides -->
                 <div class="carousel-inner">
                     <div class="item active">
-                        <img src="https://lh3.googleusercontent.com/proxy/Y3Vz3mIl0G2LulKlauT3neKxgVFL_Gyov4HrqrzcGYIeQMkNfbN_8y4N8CEXKr1sztLJ9BApjprltW8sUAWmN5N3ixifxDrC686AUo819P-8ugLtZiwxY3aLCSfaNgNkpZCMIrX_w8kf0MdZYyYIOMZAb3TSmcfEAY71jXYA761BmObdG6Ra6kf3PLpwHxB4xWUcTydgjQNbUEJonu1daXBly4TpTGW4gwos4-WdYw" alt="City of Tampa" style="width:100%;">
+                        <img src="https://www.everettpotter.com/wp-content/uploads/2010/10/tampa2.jpg" alt="City of Tampa" style="width:100%;">
+                        <div class="carousel-caption">
+                            <p>Welcome to Tampa!</p>
+                        </div>
                     </div>
 
                     <div class="item">
                         <img src="https://www.tampabay.com/resizer//d3ZAO-dvIXBQxyhlC-EhRzDqsVc=/1140x641/smart/arc-anglerfish-arc2-prod-tbt.s3.amazonaws.com/public/PKG7Y5WEREI6TKTRIBWI6S7HAY.jpg" alt="Tampa Skyline" style="width: 100%;">
+                        <div class="carousel-caption">
+                            <p>Tampa Skyline</p>
+                        </div>
                     </div>
 
                     <div class="item">
                         <img src="https://www.wfla.com/wp-content/uploads/sites/71/2020/04/bucs_1552661303448_71582599_ver1.0-7-1-1.jpg?w=900" alt="Buccaneers" style="width: 100%;">
+                        <div class="carousel-caption">
+                            <p>Buccaneers</p>
+                        </div>
                     </div>
 
                     <div class="item">
                         <img src="https://www.tampagov.net/sites/default/files/styles/slideshow_full_2016/public/web-team/images/fire_1.jpg?itok=Ews8Aqcb" alt="Tampa Fire Rescue" style="width: 100%;">
+                        <div class="carousel-caption">
+                            <p>Tampa Fire Rescue</p>
+                        </div>
                     </div>
 
                     <div class="item">
                         <img src="https://www.tampagov.net/sites/default/files/styles/slideshow_full_2016/public/web-team/images/transportation_2.jpg?itok=m37JA-3r" alt="Tampa Electric" style="width: 100%;">
+                        <div class="carousel-caption">
+                            <p>Tampa Electric</p>
+                        </div>
                     </div>
 
                     <div class="item">
                         <img src="https://www.tampagov.net/sites/default/files/styles/slideshow_full_2016/public/web-team/images/fleet_3.jpg?itok=Fh9WZXg4" alt="Tampa Engineering" style="width: 100%;">
+                        <div class="carousel-caption">
+                            <p>Tampa Engineering</p>
+                        </div>
                     </div>
 
-                    <div class="item">
-                        <img src="https://ak.picdn.net/shutterstock/videos/18348361/thumb/12.jpg" alt="Tampa Skyline at night!" style="width: 100%;">
+                <div class="item">
+                    <img src="https://ak.picdn.net/shutterstock/videos/18348361/thumb/12.jpg" alt="Tampa Skyline at night!" style="width: 100%;">
+                    <div class="carousel-caption">
+                        <p>Tampa Skyline at night!</p>
                     </div>
                 </div>
 
@@ -97,6 +117,7 @@
                     <span class="glyphicon glyphicon-chevron-right"></span>
                     <span class="sr-only">Next</span>
                 </a>
+                </div>
             </div>
         </div>
     </td>
