@@ -32,9 +32,10 @@
                 <div class="navbar-collapse collapse">
                     <ul class="nav navbar-nav">
                         <li><a asp-area="" asp-controller="Home" asp-action="Index">Home</a></li>
-                        <li><a asp-area="" asp-controller="Home" asp-action="Attractions">Attractions</a></li>
+                        <li><a asp-area="" asp-controller="Home" asp-action="Attractions"> Attractions </a></li>
                         <li><a asp-area="" asp-controller="Home" asp-action="AboutCityOfTampa"> About Tampa </a></li>
                         <li><a asp-area="" asp-controller="Home" asp-action="AboutUS"> About Us </a></li>
+                        <li><a asp-area="" asp-controller="Home" asp-action="UserLogin"> User Login </a></li>
                     </ul>
                 </div>
             </div>
